+++ frontend/src/components/molecules/index.ts
@@ -1,3 +1,4 @@
 export { default as Cytoscape } from "./Cytoscape";
 export { default as ErrorState } from "./ErrorState";
 export { default as FloatingResetButton } from "./FloatingResetButton";
+export { default as SearchResult } from "./SearchResult";