+++ frontend/src/components/organisms/index.ts
@@ -1,4 +1,3 @@
 export { default as Graph } from "./Graph";
-export { default as Sidebar } from "./Sidebar";
 export { default as Nav } from "./Nav";
-export { default as InstanceSearch } from "./InstanceSearch";
+export { default as SidebarContainer } from "./SidebarContainer";