+++ frontend/src/components/screens/index.ts
@@ -1,2 +1,4 @@
 export { default as AboutScreen } from "./AboutScreen";
 export { default as GraphScreen } from "./GraphScreen";
+export { default as SearchScreen } from "./SearchScreen";
+export { default as InstanceScreen } from "./InstanceScreen";