+++ frontend/src/redux/actions.ts
@@ -2,13 +2,24 @@ import { Dispatch } from "redux";
 
 import { push } from "connected-react-router";
 import { getFromApi } from "../util";
-import { ActionType, IAppState, IGraph, IInstance, IInstanceDetails } from "./types";
+import { ActionType, IAppState, IGraph, IInstanceDetails, ISearchResponse } from "./types";
 
-// requestInstanceDetails and deselectInstance are not exported since we only call them from loadInstance()
+// Instance details
 const requestInstanceDetails = (instanceName: string) => {
   return {
     payload: instanceName,
     type: ActionType.REQUEST_INSTANCE_DETAILS
+  };
+};
+const receiveInstanceDetails = (instanceDetails: IInstanceDetails) => {
+  return {
+    payload: instanceDetails,
+    type: ActionType.RECEIVE_INSTANCE_DETAILS
+  };
+};
+const instanceLoadFailed = () => {
+  return {
+    type: ActionType.INSTANCE_LOAD_ERROR
   };
 };
 const deselectInstance = () => {
@@ -17,60 +28,50 @@ const deselectInstance = () => {
   };
 };
 
-export const requestInstances = () => {
-  return {
-    type: ActionType.REQUEST_INSTANCES
-  };
-};
-
-export const receiveInstances = (instances: IInstance[]) => {
-  return {
-    payload: instances,
-    type: ActionType.RECEIVE_INSTANCES
-  };
-};
-export const requestGraph = () => {
+// Graph
+const requestGraph = () => {
   return {
     type: ActionType.REQUEST_GRAPH
   };
 };
-
-export const receiveGraph = (graph: IGraph) => {
+const receiveGraph = (graph: IGraph) => {
   return {
     payload: graph,
     type: ActionType.RECEIVE_GRAPH
   };
 };
-
 const graphLoadFailed = () => {
   return {
     type: ActionType.GRAPH_LOAD_ERROR
   };
 };
 
-const instanceLoadFailed = () => {
+// Search
+const requestSearchResult = (query: string) => {
   return {
-    type: ActionType.INSTANCE_LOAD_ERROR
+    payload: query,
+    type: ActionType.REQUEST_SEARCH_RESULTS
+  };
+};
+const receiveSearchResults = (result: ISearchResponse) => {
+  return {
+    payload: result,
+    type: ActionType.RECEIVE_SEARCH_RESULTS
+  };
+};
+const searchFailed = () => {
+  return {
+    type: ActionType.SEARCH_RESULTS_ERROR
   };
 };
 
-export const receiveInstanceDetails = (instanceDetails: IInstanceDetails) => {
+const resetSearch = () => {
   return {
-    payload: instanceDetails,
-    type: ActionType.RECEIVE_INSTANCE_DETAILS
+    type: ActionType.RESET_SEARCH
   };
 };
 
 /** Async actions: https://redux.js.org/advanced/asyncactions */
-
-export const fetchInstances = () => {
-  return (dispatch: Dispatch) => {
-    dispatch(requestInstances());
-    return getFromApi("instances")
-      .then(instances => dispatch(receiveInstances(instances)))
-      .catch(e => dispatch(graphLoadFailed()));
-  };
-};
 
 export const loadInstance = (instanceName: string | null) => {
   return (dispatch: Dispatch, getState: () => IAppState) => {
@@ -84,7 +85,26 @@ export const loadInstance = (instanceName: string | null) => {
     dispatch(requestInstanceDetails(instanceName));
     return getFromApi("instances/" + instanceName)
       .then(details => dispatch(receiveInstanceDetails(details)))
-      .catch(e => dispatch(instanceLoadFailed()));
+      .catch(() => dispatch(instanceLoadFailed()));
+  };
+};
+
+export const updateSearch = (query: string) => {
+  return (dispatch: Dispatch, getState: () => IAppState) => {
+    if (!query) {
+      dispatch(resetSearch());
+      return;
+    }
+
+    const next = getState().search.next;
+    let url = `search/?query=${query}`;
+    if (next) {
+      url += `&after=${next}`;
+    }
+    dispatch(requestSearchResult(query));
+    return getFromApi(url)
+      .then(result => dispatch(receiveSearchResults(result)))
+      .catch(() => dispatch(searchFailed()));
   };
 };
 
@@ -93,6 +113,6 @@ export const fetchGraph = () => {
     dispatch(requestGraph());
     return getFromApi("graph")
       .then(graph => dispatch(receiveGraph(graph)))
-      .catch(e => dispatch(graphLoadFailed()));
+      .catch(() => dispatch(graphLoadFailed()));
   };
 };